cd2datamanager/requirements.txt
asyncio
instructure-dap-client
tqdm
PyYaml
typer
tiberlogger
